pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "iref"
version = "0.0.1"
description = "Illustration Refs – Discovery Queue CLI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "typer[all]",
  "rich",
  "pillow",
  "ImageHash",
  "watchdog",
  "orjson",
  "python-dateutil",
  "platformdirs",
  "readchar>=4.0",          # ← 追加
]

[project.scripts]
iref = "iref.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/iref"]   # ← ここをこの1行に。ほかは不要